# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Auditoria do Condomínio – API
  version: 0.1.0
  description: |
    Contrato entre backend, frontend e MCP. Toda chamada exige token Bearer do Keycloak.
    Perfis: USER (consulta), MANAGER (+ envia arquivos), ADMIN (tudo).
    Valores monetários são números com 2 casas decimais; datas em ISO 8601.
servers:
  - url: /api
security:
  - bearer: []
paths:
...
  /condominiums/{condominiumId}/dashboard:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    get:
      summary: Números da tela inicial, do fluxo de caixa mais recente
      operationId: dashboard
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DashboardResponse" } } } }
        "204": { description: Ainda não há fluxo de caixa processado }
  /condominiums/{condominiumId}/operating-fund:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    put:
      summary: Confirma qual fundo é o ordinário do condomínio (MANAGER ou ADMIN)
      operationId: setOperatingFund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fundId]
              properties: { fundId: { type: string, format: uuid } }
      responses:
        "204": { description: Confirmado }
        "403": { description: Perfil sem permissão }
        "404": { description: Fundo não é deste condomínio }
...
  /condominiums/{condominiumId}/budget-vs-actual:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    get:
      summary: Previsto × realizado do mês ou acumulado do exercício (todos os perfis, RF-03.1.6 a RF-03.1.11)
      description: |
        Calculado na consulta pela função pura CalculoPrevistoRealizado (versaoCalculo no resultado); nada é gravado.
        Realizado = débitos do fundo Condomínio (fundo ordinário) pelo de-para CONFIRMED; ajustes, a realocar e sem
        linha da PO à parte; transferências entre fundos fora. Previsto do mês = soma das linhas 1.1 a 1.8 (Q30).
        Fundos ligados às linhas 1.9: arrecadação (recebimento de cota) × previsto. Mês sem fluxo, com dois fluxos,
        sem PO, com PO não confirmada ou sem fundo ordinário: situacao diz o porquê e os números ficam nulos.
      operationId: budgetVsActual
      parameters:
        - { name: period, in: query, required: true, description: AAAA-MM ou "cumulative", schema: { type: string, pattern: "^([0-9]{4}-[0-9]{2}|cumulative)$" } }
        - { name: budget, in: query, required: false, description: Versão da PO; sem ela, a que vale no mês (no acumulado, a confirmada mais recente), schema: { type: string, format: uuid } }
        - { name: fund, in: query, required: false, description: 'Filtro de fundo (RF-03.1.13). O fundo ordinário mostra só o fundo Condomínio (sem o painel dos fundos); outro fundo mostra só o painel dele, com totais, grupos, blocos, conferência e regra dos 20% nulos ou vazios e provisorio = false. Sem ele, tudo', schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/BudgetVsActualResponse" } } } }
        "400": { description: Período inválido }
        "403": { description: Sem acesso ao condomínio }
        "404": { description: Condomínio, PO ou fundo não encontrados }
  /condominiums/{condominiumId}/budget-vs-actual/evidence:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    get:
      summary: Lançamentos que compõem um número do previsto × realizado (todos os perfis, RF-03.1.12)
      operationId: budgetVsActualEvidence
      parameters:
        - { name: period, in: query, required: true, schema: { type: string, pattern: "^([0-9]{4}-[0-9]{2}|cumulative)$" } }
        - { name: budget, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: target, in: query, required: true, description: '"line:<lineId>", "group:<lineId do grupo>" (as linhas do grupo, na ordem da PO), "total" (despesa realizada: linhas de todos os grupos, depois TO_REALLOCATE e NO_BUDGET_LINE), "fund:<fundoId>" (arrecadação), ADJUSTMENTS, TO_REALLOCATE, NO_BUDGET_LINE ou TRANSFERS', schema: { type: string } }
      responses:
        "200":
          description: OK (lista vazia quando o alvo não tem lançamento)
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/EvidenceResponse" } } } }
        "400": { description: Período ou alvo inválido }
        "403": { description: Sem acesso ao condomínio }
...
  /condominiums/{condominiumId}/assistant/questions:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    post:
      summary: Pergunta ao assistente, com citações dos documentos e números das ferramentas (USER, MANAGER ou ADMIN; RF-04.8 a RF-04.16)
      description: |
        Só com o módulo Assistente ligado e o modo de respostas efetivo API_KEY. O backend chama o rag (gRPC Perguntar,
        prazo de 120 s), junta o fluxo e devolve a resposta inteira, já validada. O histórico só existe na tela
        (RF-04.11): o frontend manda as trocas anteriores e o backend usa as últimas 6. Citações de arquivos que o
        usuário não pode ver ou que foram excluídos são descartadas na volta. Gera o registro de uso "pergunta".
      operationId: askAssistant
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/QuestionRequest" } } }
      responses:
        "200": { description: Resposta (ANSWERED ou NOT_FOUND), content: { application/json: { schema: { $ref: "#/components/schemas/AssistantAnswerResponse" } } } }
        "400": { description: Pergunta vazia ou acima de 2000 caracteres, ou filtro inválido, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "403": { description: Perfil sem permissão, sem acesso ao condomínio ou módulo não contratado (modulo = ASSISTANT), content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "404": { description: Condomínio não encontrado }
        "409":
          description: |
            Chat indisponível no modo atual, sem chamar o rag (modoIa diz qual). EXTERNAL_MCP: "O assistente deste
            condomínio é o seu Claude, conectado ao MCP." OFF: "A IA está desligada neste condomínio." Também
            sem chave cadastrada ou chave que o rag não consegue ler ("cadastre a chave de novo").
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
        "422": { description: Chave de IA do condomínio recusada pelo provedor, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "429": { description: Limite de uso do provedor de IA atingido; tentar mais tarde, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "503": { description: Rag ou provedor de IA fora do ar, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "504": { description: A resposta passou do prazo, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  /condominiums/{condominiumId}/assistant/search:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    post:
      summary: Busca por palavra nos documentos, sem IA (todos os perfis; RF-04.18)
      description: |
        Só com o módulo Assistente ligado, em qualquer modo de IA (inclusive OFF, Q7). Devolve trechos citáveis
        do mais relevante para o menos relevante, sem resposta redigida. Aceita "frase entre aspas" e exclusão com -.
        Gera o registro de uso "busca_documentos".
      operationId: searchDocuments
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DocumentSearchRequest" } } }
      responses:
        "200":
          description: OK (lista vazia quando nada foi encontrado)
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/DocumentChunkResponse" } } } }
        "400": { description: Texto vazio, limite fora de 1 a 50 ou filtro inválido, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "403": { description: Sem acesso ao condomínio ou módulo não contratado (modulo = ASSISTANT), content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "404": { description: Condomínio não encontrado }
        "503": { description: Rag fora do ar, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  /condominiums/{condominiumId}/budget-vs-actual/export:
    parameters: [{ $ref: "#/components/parameters/CondominiumId" }]
    get:
      summary: Exporta a mesma visão do previsto × realizado em PDF ou Excel (todos os perfis, RF-03.1.14)
      description: |
        Gerado do mesmo resultado do GET /previsto-realizado (mesma função, mesmos filtros): os números são os do
        JSON. Cabeçalho com condomínio, PO (arquivo, versão, hash e exercício), período, data e hora, quem gerou e
        estado do de-para. Com valor a realocar ou sem linha da PO, traz "PROVISÓRIO" e a lista desses lançamentos.
        Sem gráfico. Excel com as abas "Resumo" e "Evidência" (um lançamento por linha, com arquivo, página e hash).
        Mês sem números (sem fluxo, dois fluxos etc.) exporta a mensagem no lugar dos números.
      operationId: exportBudgetVsActual
      parameters:
        - { name: format, in: query, required: true, schema: { type: string, enum: [pdf, xlsx] } }
        - { name: period, in: query, required: true, description: AAAA-MM ou "cumulative", schema: { type: string, pattern: "^([0-9]{4}-[0-9]{2}|cumulative)$" } }
        - { name: budget, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: fund, in: query, required: false, description: 'Filtro de fundo (RF-03.1.13). O fundo ordinário mostra só o fundo Condomínio (sem o painel dos fundos); outro fundo mostra só o painel dele, com totais, grupos, blocos, conferência e regra dos 20% nulos ou vazios e provisorio = false. Sem ele, tudo', schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Arquivo (Content-Disposition attachment; previsto-realizado-<periodo>.pdf ou .xlsx)
          content:
            application/pdf: { schema: { type: string, format: binary } }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: { schema: { type: string, format: binary } }
        "400": { description: Formato ou período inválido }
        "403": { description: Sem acesso ao condomínio }
        "404": { description: Condomínio, PO ou fundo não encontrados }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    CondominiumId: { name: condominiumId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    QuestionRequest:
      type: object
      required: [question]
      properties:
        question: { type: string, minLength: 1, maxLength: 2000 }
        history:
          type: array
          description: Trocas anteriores desta conversa, da mais antiga para a mais recente (o backend usa as últimas 6)
          items:
            type: object
            required: [question, answer]
            properties: { question: { type: string }, answer: { type: string } }
        filters: { $ref: "#/components/schemas/DocumentFiltersRequest" }
...
    AssistantAnswerResponse:
      type: object
      required: [status, fromDocuments, fromStoredData, citations, model]
      properties:
        status: { $ref: "#/components/schemas/AnswerStatus" }
        fromDocuments:
          type: array
          description: Bloco "Nos documentos". Números transcritos vêm marcados "(conforme o documento, não conferido)"
          items:
            type: object
            required: [text, citations]
            properties:
              text: { type: string }
              citations: { type: array, items: { type: integer }, description: Números das citações (campo numero de citacoes) }
        fromStoredData:
          type: array
          description: Bloco "Nos dados gravados", montado das ferramentas de consulta (nunca escrito pelo modelo)
          items: { $ref: "#/components/schemas/StoredDataResponse" }
        citations: { type: array, items: { $ref: "#/components/schemas/DocumentCitationResponse" }, description: "Numeradas a partir de 1, na ordem da primeira citação" }
        suggestion: { type: [string, "null"], description: Categoria de documento que faltaria (RF-04.12) }
        warning: { type: [string, "null"] }
        model: { type: string, description: Modelo de IA que respondeu }
...
    DocumentChunkResponse:
      type: object
      required: [chunkId, fileId, fileName, category, location, text, sha256]
      properties:
        chunkId: { type: string }
        fileId: { type: string, format: uuid }
        fileName: { type: string }
        category: { $ref: "#/components/schemas/FileCategory" }
        location: { $ref: "#/components/schemas/ChunkLocationResponse" }
        text: { type: string, description: Texto literal do documento, não conferido }
        sha256: { type: string }
...
    DocumentSearchRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 500 }
        filters: { $ref: "#/components/schemas/DocumentFiltersRequest" }
        limit: { type: integer, minimum: 1, maximum: 50, default: 10 }
...
    DashboardResponse:
      type: object
      required: [fileId, fileName, periodStart, periodEnd, openingBalance, inflows, outflows, closingBalance, failedChecks, operatingFund, funds, largestExpenses]
      properties:
        fileId: { type: string, format: uuid }
        fileName: { type: string }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        openingBalance: { type: number }
        inflows: { type: number }
        outflows: { type: number }
        closingBalance: { type: number }
        failedChecks: { type: integer }
        operatingFund:
          description: |
            Saldo acumulado do fundo ordinário (cartão "Saldo acumulado"). Com confirmado=false é só a sugestão
            (fundo com mais entradas no mês), que a tela mostra para o Gestor confirmar. Nulo quando não há sugestão.
          oneOf:
            - type: "null"
            - type: object
              required: [fundId, fund, confirmed, closingBalance]
              properties:
                fundId: { type: string, format: uuid }
                fund: { type: string }
                confirmed: { type: boolean }
                closingBalance: { type: [number, "null"], description: Nulo quando o fundo confirmado não está no relatório }
        funds:
          type: array
          items:
            type: object
            required: [fundId, fund, openingBalance, inflows, outflows, result, closingBalance]
            properties:
              fundId: { type: string, format: uuid }
              fund: { type: string }
              openingBalance: { type: number }
              inflows: { type: number }
              outflows: { type: number }
              result: { type: number }
              closingBalance: { type: number }
        largestExpenses:
          type: array
          items:
            type: object
            required: [date, fund, account, memo, amount, page]
            properties:
              date: { type: string, format: date }
              fund: { type: string }
              account: { type: string }
              memo: { type: string }
              amount: { type: number }
              page: { type: integer }
    BudgetStatus:
      type: string
      description: |
        READ (aguarda confirmação), READ_WITH_DISCREPANCY (soma que não bate além da tolerância de arredondamento),
        CONFIRMED (vale para os meses do exercício) e SUPERSEDED (trocada por reaprovação a partir de um mês).
      enum: [READ, READ_WITH_DISCREPANCY, CONFIRMED, SUPERSEDED]
...
    BudgetVsActualResponse:
      type: object
      required: [calculationVersion, period, status, months, summedMonths, monthsWithoutCashFlow, monthsWithTwoCashFlows, provisional, groups, funds, warnings]
      properties:
        calculationVersion: { type: string, description: Versão das regras do cálculo usada nesta resposta }
        period: { type: string, description: AAAA-MM ou "cumulative" }
        status: { type: string, enum: [CALCULATED, NO_BUDGET, BUDGET_NOT_CONFIRMED, NO_OPERATING_FUND, NO_CASH_FLOW, TWO_CASH_FLOWS] }
        message: { type: [string, "null"], description: Por que não há números (situação diferente de CALCULATED) }
        budget:
          oneOf:
            - type: "null"
            - type: object
              required: [id, status, fileId, sha256]
              properties:
                id: { type: string, format: uuid }
                version: { type: [integer, "null"] }
                status: { $ref: "#/components/schemas/BudgetStatus" }
                fileId: { type: string, format: uuid }
                fileName: { type: [string, "null"] }
                sha256: { type: string }
                fiscalYearStart: { type: [string, "null"] }
                fiscalYearEnd: { type: [string, "null"] }
        months:
          type: array
          description: |
            No acumulado, os meses do exercício e, depois deles, os meses prorrogados (fora da soma, RF-11.3); no mês,
            só ele. Números só com WITH_CASH_FLOW
          items:
            type: object
            required: [month, status, cashFlows]
            properties:
              month: { type: string }
              status: { type: string, enum: [WITH_CASH_FLOW, NO_CASH_FLOW, TWO_CASH_FLOWS] }
              cashFlows: { type: array, items: { $ref: "#/components/schemas/UsedCashFlowResponse" } }
              planned: { type: [number, "null"] }
              actualExpense: { type: [number, "null"] }
              overrun: { type: [number, "null"] }
              overrunPercentage: { type: [number, "null"] }
              aboveLimit: { type: [boolean, "null"] }
              extended: { type: boolean, description: Mês depois do exercício em que a PO vale por prorrogação (marca "PO prorrogada") }
        summedMonths: { type: array, items: { type: string } }
        monthsWithoutCashFlow: { type: array, items: { type: string }, description: Meses sem fluxo antes do último mês com fluxo }
        monthsWithTwoCashFlows: { type: array, items: { type: string } }
        mapping:
          oneOf:
            - type: "null"
            - type: object
              required: [accounts, confirmed, withoutConfirmedMapping]
              properties:
                accounts: { type: integer, description: Contas com débito no fundo Condomínio no período }
                confirmed: { type: integer }
                withoutConfirmedMapping: { type: integer }
        provisional: { type: boolean, description: Há valor a realocar ou sem linha da PO }
        totals:
          oneOf:
            - type: "null"
            - type: object
              required: [monthlyPlanned, planned, actualExpense, inLines, difference, fiscalYearPlanned]
              properties:
                monthlyPlanned: { type: number }
                planned: { type: number, description: Previsto do mês × meses somados }
                actualExpense: { type: number, description: Em linhas + a realocar + sem linha da PO }
                inLines: { type: number }
                difference: { type: number }
                execution: { type: [number, "null"], description: "% com 1 casa; nulo com previsto zero" }
                fiscalYearPlanned: { type: number, description: Previsto do mês × meses do exercício (referência) }
        groups: { type: array, items: { $ref: "#/components/schemas/BudgetVsActualGroupResponse" } }
        adjustments: { oneOf: [{ type: "null" }, { $ref: "#/components/schemas/EntryBlockResponse" }] }
        toReallocate: { oneOf: [{ type: "null" }, { $ref: "#/components/schemas/EntryBlockResponse" }] }
        withoutBudgetLine: { oneOf: [{ type: "null" }, { $ref: "#/components/schemas/EntryBlockResponse" }] }
        cashFlowCheck:
          oneOf:
            - type: "null"
            - type: object
              description: Total de débitos do fundo = despesa realizada + ajustes + transferências
              required: [fundDebits, entries, actualExpense, adjustments, transfers, matches]
              properties:
                fundDebits: { type: number }
                entries: { type: integer }
                actualExpense: { type: number }
                adjustments: { type: number }
                transfers: { type: number }
                matches: { type: boolean }
        rule20:
          description: Regra dos 20% (Conv. 16.2), só no mês
          oneOf:
            - type: "null"
            - type: object
              required: [rule, ruleVersion, limitPercentage, monthlyPlanned, overrun, percentage, limit, linesAbove, lines, toReallocate, withoutBudgetLine, maxScenario, provisional, aboveLimit]
              properties:
                rule: { type: string }
                ruleVersion: { type: string }
                limitPercentage: { type: number }
                monthlyPlanned: { type: number }
                overrun: { type: number, description: Soma das diferenças positivas, linha a linha }
                percentage: { type: number }
                limit: { type: number, description: Em reais, arredondado; a comparação é exata }
                linesAbove: { type: integer }
                lines:
                  type: array
                  items:
                    type: object
                    required: [lineId, code, description, overrun]
                    properties:
                      lineId: { type: string, format: uuid }
                      code: { type: string }
                      description: { type: string }
                      overrun: { type: number }
                toReallocate: { type: number }
                withoutBudgetLine: { type: number }
                maxScenario: { type: number, description: Excesso + a realocar + sem linha da PO }
                maxScenarioPercentage: { type: [number, "null"] }
                provisional: { type: boolean }
                aboveLimit: { type: boolean }
        funds:
          type: array
          items:
            type: object
            required: [status]
            properties:
              fundId: { type: [string, "null"], format: uuid }
              fund: { type: [string, "null"], description: Nome impresso no fluxo }
              lineId: { type: [string, "null"], format: uuid }
              lineCode: { type: [string, "null"] }
              status: { type: string, enum: [COMPARED, NOT_PLANNED_IN_BUDGET, LINE_WITHOUT_FUND, REPROCESS_CASH_FLOW] }
              planned: { type: [number, "null"] }
              collected: { type: [number, "null"] }
              difference: { type: [number, "null"] }
              execution: { type: [number, "null"] }
              credits: { type: [number, "null"] }
              debits: { type: [number, "null"] }
        warnings:
          type: array
          description: Inclui BUDGET_EXTENDED (mês prorrogado) e EXTENDED_MONTHS (acumulado com meses fora da soma)
          items:
            type: object
            required: [code, text]
            properties: { code: { type: string }, text: { type: string } }
    UsedCashFlowResponse:
      type: object
      required: [fileId, name, sha256]
      properties:
        fileId: { type: string, format: uuid }
        name: { type: string }
        sha256: { type: string }
        periodStart: { type: [string, "null"], format: date }
        periodEnd: { type: [string, "null"], format: date }
        uploadedAt: { type: [string, "null"], format: date-time }
        uploadedBy: { type: [string, "null"] }
    BudgetVsActualGroupResponse:
      type: object
      required: [lineId, code, description, planned, actual, difference, lines]
      properties:
        lineId: { type: string, format: uuid }
        code: { type: string }
        description: { type: string }
        planned: { type: number, description: Soma das linhas (Q30) }
        actual: { type: number }
        difference: { type: number }
        execution: { type: [number, "null"] }
        lines:
          type: array
          items:
            type: object
            required: [lineId, code, description, page, planned, actual, difference, cashFlowAccounts, entries]
            properties:
              lineId: { type: string, format: uuid }
              code: { type: string }
              description: { type: string }
              account: { type: [string, "null"] }
              mark: { type: [string, "null"], enum: [SEPARATE_APPORTIONMENT, NEGOTIATED_EXEMPTION, NO_AMOUNT, FIXED_AMOUNT_NO_REFERENCE, null] }
              notes: { type: [string, "null"] }
              page: { type: integer }
              planned: { type: number }
              actual: { type: number }
              difference: { type: number }
              execution: { type: [number, "null"] }
              cashFlowAccounts: { type: array, items: { type: string }, description: Contas do fluxo com de-para confirmado para a linha }
              entries: { type: integer }
...
    EvidenceResponse:
      type: object
      required: [entryId, date, memo, amount, fileId, sha256, page, position]
      properties:
        entryId: { type: string, format: uuid }
        date: { type: string, format: date }
        account: { type: [string, "null"] }
        accountName: { type: [string, "null"] }
        memo: { type: string }
        supplier: { type: [string, "null"] }
        document: { type: [string, "null"] }
        amount: { type: number }
        fund: { type: [string, "null"] }
        fileId: { type: string, format: uuid }
        fileName: { type: [string, "null"] }
        sha256: { type: string }
        page: { type: integer }
        position: { type: integer }
        reallocation: { type: [string, "null"], description: '"realocado para <linha> por <usuário> em <data>"' }
        reallocationId: { type: [string, "null"], format: uuid, description: Realocação ativa que levou o lançamento à linha (para desfazer) }
...
    Problem:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        existingFileId: { type: string, format: uuid }
        feature: { type: string, description: "Em recusa por módulo não contratado (403): código do módulo" }
        reasons: { type: array, items: { type: string }, description: "Motivos da recusa (confirmação da PO, configuração de IA)" }
        aiMode: { $ref: "#/components/schemas/AiMode", description: "Em recusa do chat pelo modo (409): modo de respostas efetivo" }